[Imp] Show sample numbers in Sample Map editor (http://forum.openmpt.org/index.php?topic=1795.msg15678#msg15678)
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@5659 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/PatternEditorDialogs.cpp
@@ -1340,7 +1340,7 @@ void CChordEditor::UpdateKeyboard()
 	}
 	for(UINT i = 0; i < 2 * 12; i++)
 	{
-		UINT b = CKeyboardControl::KEYFLAG_NORMAL;
+		uint8 b = CKeyboardControl::KEYFLAG_NORMAL;
 		if(i == note) b = CKeyboardControl::KEYFLAG_REDDOT;
 		else if(chord.notes[0] && i + 1 == chord.notes[0]) b = CKeyboardControl::KEYFLAG_REDDOT;
 		else if(chord.notes[1] && i + 1 == chord.notes[1]) b = CKeyboardControl::KEYFLAG_REDDOT;
